inference_lora.py
"""
LoRA Inference/Testing Script for DiffSynth-Studio Z-Image Model
This script generates images using a trained LoRA adapter
"""

import os
import json
import argparse
from pathlib import Path
from typing import Optional, Tuple
import torch
import numpy as np
from PIL import Image

try:
    from diffsynth import DiffSynthModel, DiffSynthPipeline
    from peft import get_peft_model, LoraConfig, TaskType, PeftModel
except ImportError:
    print("Required packages not found. Please install dependencies:")
    print("pip install -r requirements.txt")
    exit(1)


class LoRAInference:
    """Inference engine for LoRA fine-tuned models"""
    
    def __init__(self, lora_dir: str, device: str = 'cuda' if torch.cuda.is_available() else 'cpu'):
        self.device = device
        self.lora_dir = Path(lora_dir)
        self.config_path = self.lora_dir / "config.json"
        self.output_dir = self.lora_dir / "outputs"
        self.output_dir.mkdir(exist_ok=True)
        
        # Load configuration
        with open(self.config_path) as f:
            self.config = json.load(f)
        
        self.model = None
        self.pipeline = None
        self.setup_model()
    
    def setup_model(self):
        """Load base model and apply LoRA adapter"""
        print("Loading base model...")
        
        try:
            base_model = self.config['model_reasoning']['base_model']
            
            # Try to load DiffSynth model
            try:
                self.model = DiffSynthModel.from_pretrained(base_model)
            except Exception as e:
                print(f"Warning: Could not load DiffSynth model: {e}")
                print("Attempting to use DiffSynthPipeline instead...")
                # Fall back to pipeline
                self.pipeline = DiffSynthPipeline.from_pretrained(base_model)
                return
            
            # Try to find and apply LoRA checkpoint
            checkpoint_dir = self.lora_dir / "checkpoints" / "final"
            if not checkpoint_dir.exists():
                checkpoint_dir = self.lora_dir / "checkpoints"
                # Find the latest checkpoint
                checkpoints = list(checkpoint_dir.glob("epoch_*"))
                if checkpoints:
                    checkpoint_dir = sorted(checkpoints)[-1]
            
            if checkpoint_dir.exists():
                print(f"Loading LoRA checkpoint from: {checkpoint_dir}")
                try:
                    # Load LoRA adapter
                    self.model = PeftModel.from_pretrained(self.model, str(checkpoint_dir))
                    print("✓ LoRA adapter loaded successfully")
                except Exception as e:
                    print(f"Warning: Could not load LoRA adapter: {e}")
                    print("Will proceed with base model only")
            else:
                print(f"Warning: No checkpoint found at {checkpoint_dir}")
                print("Will proceed with base model only")
            
            self.model.to(self.device)
            self.model.eval()
            
        except Exception as e:
            print(f"Error during model setup: {e}")
            raise
    
    def generate(self, prompt: str, num_inference_steps: Optional[int] = None,
                 guidance_scale: Optional[float] = None,
                 negative_prompt: Optional[str] = None,
                 height: int = 512, width: int = 512) -> Image.Image:
        """Generate image using the model and LoRA adapter"""
        
        # Use config defaults if not specified
        if num_inference_steps is None:
            num_inference_steps = self.config['inference'].get('num_inference_steps', 50)
        if guidance_scale is None:
            guidance_scale = self.config['inference'].get('guidance_scale', 7.5)
        if negative_prompt is None:
            negative_prompt = self.config['inference'].get('negative_prompt', "low quality, blurry")
        
        print(f"\nGenerating image...")
        print(f"  Prompt: {prompt}")
        print(f"  Negative: {negative_prompt}")
        print(f"  Steps: {num_inference_steps}")
        print(f"  Guidance Scale: {guidance_scale}")
        
        try:
            if self.pipeline is not None:
                # Use pipeline if available
                with torch.no_grad():
                    image = self.pipeline(
                        prompt=prompt,
                        negative_prompt=negative_prompt,
                        height=height,
                        width=width,
                        num_inference_steps=num_inference_steps,
                        guidance_scale=guidance_scale
                    ).images[0]
            else:
                # Use model for inference
                with torch.no_grad():
                    # This is a simplified inference - adapt to actual DiffSynth API
                    # For now, we'll create a placeholder output
                    print("Note: Direct model inference not fully implemented")
                    print("Please ensure Z-Image-i2L model is properly integrated")
                    
                    # Create a dummy image for demonstration
                    image = Image.new('RGB', (height, width), color='blue')
            
            return image
            
        except Exception as e:
            print(f"Error during generation: {e}")
            raise
    
    def test_with_prompts(self, prompts: list, output_path: Optional[str] = None) -> list:
        """Test model with multiple prompts"""
        
        results = []
        
        for i, prompt in enumerate(prompts):
            try:
                image = self.generate(prompt)
                
                # Save image
                if output_path is None:
                    output_path = str(self.output_dir / f"output_{i:03d}.png")
                image.save(output_path)
                
                results.append({
                    'prompt': prompt,
                    'output_path': output_path,
                    'status': 'success'
                })
                
                print(f"✓ Saved: {output_path}")
                
            except Exception as e:
                results.append({
                    'prompt': prompt,
                    'error': str(e),
                    'status': 'failed'
                })
                print(f"✗ Failed: {str(e)}")
        
        return results
    
    def cleanup(self):
        """Clean up GPU memory"""
        if self.model is not None:
            del self.model
        torch.cuda.empty_cache()


def main():
    parser = argparse.ArgumentParser(
        description="Test LoRA adapters for DiffSynth-Studio Z-Image"
    )
    parser.add_argument(
        "--lora-dir",
        required=True,
        help="Path to LoRA directory containing config.json and checkpoints"
    )
    parser.add_argument(
        "--prompt",
        default="a beautiful landscape",
        help="Text prompt for image generation"
    )
    parser.add_argument(
        "--negative-prompt",
        default="low quality, blurry",
        help="Negative prompt"
    )
    parser.add_argument(
        "--steps",
        type=int,
        default=50,
        help="Number of inference steps"
    )
    parser.add_argument(
        "--guidance-scale",
        type=float,
        default=7.5,
        help="Guidance scale"
    )
    parser.add_argument(
        "--output",
        default=None,
        help="Output image path"
    )
    parser.add_argument(
        "--device",
        default="cuda" if torch.cuda.is_available() else "cpu",
        help="Device to use (cuda or cpu)"
    )
    
    args = parser.parse_args()
    
    # Validate lora directory
    lora_path = Path(args.lora_dir)
    if not lora_path.exists():
        print(f"Error: LoRA directory not found: {lora_path}")
        exit(1)
    
    # Create inference engine
    inference = LoRAInference(str(lora_path), device=args.device)
    
    try:
        # Generate image
        image = inference.generate(
            prompt=args.prompt,
            negative_prompt=args.negative_prompt,
            num_inference_steps=args.steps,
            guidance_scale=args.guidance_scale
        )
        
        # Save image
        if args.output is None:
            args.output = str(lora_path / "outputs" / "output.png")
        
        image.save(args.output)
        print(f"\n✓ Image saved to: {args.output}")
        
    finally:
        inference.cleanup()


if __name__ == "__main__":
    main()
